Flask/requirements.txt
flask
pillow
torch
yolov5
gunicorn